Move Subthought into a separate file.
